Added a sub-menus and a menu item.
Consider this as a full fledge example on how to append new parameters to menu systems.

Nota that this is a first implementation of the menu, and there is plenty of room to improve and to make the operation to add submenus or menuitems a lot more easier.

TIP -> ie: Using macros! etc